+++ swap/src/asb/command.rs
@@ -31,12 +31,12 @@ where
             env_config: env_config(testnet),
             cmd: Command::Start { resume_only },
         },
-        RawCommand::History => Arguments {
+        RawCommand::History { only_unfinished } => Arguments {
             testnet,
             json,
             config_path: config_path(config, testnet)?,
             env_config: env_config(testnet),
-            cmd: Command::History,
+            cmd: Command::History { only_unfinished },
         },
         RawCommand::Logs {
             logs_dir: dir_path,
@@ -197,7 +197,9 @@ pub enum Command {
     Start {
         resume_only: bool,
     },
-    History,
+    History {
+        only_unfinished: bool,
+    },
     Config,
     Logs {
         logs_dir: Option<PathBuf>,
@@ -295,7 +297,10 @@ pub enum RawCommand {
         swap_id: Option<Uuid>,
     },
     #[structopt(about = "Prints swap-id and the state of each swap ever made.")]
-    History,
+    History {
+        #[structopt(long = "only-unfinished", help = "Only print in progress swaps")]
+        only_unfinished: bool,
+    },
     #[structopt(about = "Prints the current config")]
     Config,
     #[structopt(about = "Allows withdrawing BTC from the internal Bitcoin wallet.")]
@@ -411,7 +416,9 @@ mod tests {
             json: false,
             config_path: default_mainnet_conf_path,
             env_config: mainnet_env_config,
-            cmd: Command::History,
+            cmd: Command::History {
+                only_unfinished: false,
+            },
         };
         let args = parse_args(raw_ars).unwrap();
         assert_eq!(expected_args, args);
@@ -586,7 +593,9 @@ mod tests {
             json: false,
             config_path: default_testnet_conf_path,
             env_config: testnet_env_config,
-            cmd: Command::History,
+            cmd: Command::History {
+                only_unfinished: false,
+            },
         };
         let args = parse_args(raw_ars).unwrap();
         assert_eq!(expected_args, args);